Replaced the FOUND_* macros by static functions. Changed the found_data from a heap buffer to a fixed-size static buffer. Use size_t instead of gint where appropriate. Added assertions.

+++ src/core/qhit.h
@@ -30,7 +30,7 @@
 
 #include <glib.h>
 
-typedef void (*qhit_process_t)(gpointer data, gint len, gpointer udata);
+typedef void (*qhit_process_t)(gpointer data, size_t len, gpointer udata);
 
 /*
  * Public interface.
